created ChunkManager, moved a bunch of methods out of the Utilities class and changed references accordingly
